Add files via upload
A number game for two players. There is a random, user, and strategic player in this game. A count starts at 0. On a player's turn, they add to the count an amount between a set minimum and a set maximum. The player who brings the count to a set goal amount is the winner. The game can have multiple rounds.